Refactor code to move some utility functions into a utilities module.
